express error handler, setup page, image upload with image compression, server start/stop script

+++ storage_ram.js
@@ -1,12 +1,12 @@
 import { Content } from './content.js';
-import { rename, unlink, writeFile } from 'node:fs/promises';
+import { rename, unlink, writeFile, access, constants } from 'node:fs/promises';
 let content;
 let notes; // readonly
 /** store content in RAM */
 export function storeContent(c) {
     console.log("store content in RAM");
     // prüüüüüüüüüüüüüfen
-    if ((c instanceof Content)) {
+    if ((c instanceof Content)) { // unnötig? prüfen ob man's weglassen kann
         content = c;
     }
     else {
@@ -94,18 +94,48 @@ export function setFilename(name) {
     filenameLock = filename + ".lock";
 }
 /** write everything to file */
+// export async function writeToFile(req: Request, res: Response, next: NextFunction) { //async
+/*export async function writeToFile() { //async
+    storeContent(content) // sync
+    
+    const json = JSON.stringify(content)
+    console.log("write content to file")
+    try {
+        await rename(filename, filenameLock) // lock file // async
+        await writeFile(filename, json) // write // async
+        await unlink(filenameLock) // remove lock file //
+        console.log("written to file")
+    } catch(error) {
+        console.error(error)
+        console.log("write to file failed!")
+        //next(error)
+    }
+}*/
+/** write everything to file */
+// export async function writeToFile(req: Request, res: Response, next: NextFunction) { //async
+// pass errors to model
 export async function writeToFile() {
     storeContent(content); // sync
     const json = JSON.stringify(content);
     console.log("write content to file");
+    //try {
+    let exists = false;
     try {
+        await access('config/content.json', constants.R_OK | constants.W_OK);
+        exists = true;
+    }
+    catch (err) {
+        console.error('cannot access config/content.json');
+    }
+    if (exists)
         await rename(filename, filenameLock); // lock file // async
-        await writeFile(filename, json); // write // async
+    await writeFile(filename, json); // write // async
+    if (exists)
         await unlink(filenameLock); // remove lock file //
-        console.log("written to file");
-    }
-    catch (error) {
-        console.error(error);
-        console.log("write to file failed!");
-    }
+    console.log("written to file");
+    //} catch(error) {
+    //console.error(error)
+    //console.log("write to file failed!");
+    //next(error)
+    //}
 }
